Implmentado autenticação e token mais controles de acesso

+++ src/main/java/io/lrsystem/ServiceLog/service/AtendimentoService.java
@@ -8,10 +8,13 @@ import io.lrsystem.ServiceLog.model.Atendimento;
 import io.lrsystem.ServiceLog.model.Usuario;
 import io.lrsystem.ServiceLog.repository.AtendimentoRepository;
 import io.lrsystem.ServiceLog.repository.UsuarioRepository;
+import io.lrsystem.ServiceLog.service.exceptions.ForbiddenException;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,6 +24,7 @@ import java.time.LocalTime;
 import java.time.OffsetDateTime;
 import java.time.format.TextStyle;
 import java.util.Locale;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -29,32 +33,36 @@ public class AtendimentoService {
     private final AtendimentoMapper atendimentoMapper;
     private final AtendimentoRepository atendimentoRepository;
     private final UsuarioRepository usuarioRepository;
+    private  final AuthService authService;
 
     @Transactional(readOnly = true)
-    public Page<AtendimentoResponseDTO> listarPorUsuario(Long usuarioId,
-                                                         LocalDate inicio,
+    public Page<AtendimentoResponseDTO> listarPorUsuario(LocalDate inicio,
                                                          LocalDate fim,
                                                          Pageable pageable) {
+
+        Usuario usuario = authService.authenticated();
 
         if (inicio != null && fim != null && fim.isBefore(inicio)){
             throw new IllegalArgumentException("Data fim não pode ser menor que data inicio");
         }
 
         Page<Atendimento> atendimentos =
-                atendimentoRepository.buscarPorUsuarioEPeriodo(usuarioId,inicio,fim,pageable);
+                atendimentoRepository.buscarPorUsuarioEPeriodo(usuario.getId(),inicio,fim,pageable);
         return atendimentoMapper.toDoList(atendimentos);
     }
 
     @Transactional
     public AtendimentoResponseDTO adicionar(AtendimentoRequestDTO dto) {
 
-        Usuario usuario = usuarioRepository.findById(dto.getUsuarioId())
+        Usuario user = authService.authenticated();
+
+        Usuario usuario = usuarioRepository.findById(user.getId())
                 .orElseThrow(() -> new EntityNotFoundException("Usuario não encontrado"));
 
         Atendimento atendimento = atendimentoMapper.toEntity(dto);
         atendimento.setUsuario(usuario);
 
-        atendimento.setColaborador(usuario.getNome()); // PEGAR O USUARIO LOGADO NO SISTEMA.
+        atendimento.setColaborador(usuario.getNome());
 
         atendimento.setTempoTotal(calcularTempoTotal(atendimento.getInicio(),atendimento.getFim()));
 
@@ -85,6 +93,13 @@ public class AtendimentoService {
 
     @Transactional
     public void deletar(Long id) {
+
+        var atendimento = atendimentoRepository.findById(id).orElseThrow(
+                () -> new AtendimentoNaoEncontrado("Atendimento não encontrado")
+        );
+
+        authService.validateSelfOrAdmin(atendimento.getUsuario().getId());
+
         atendimentoRepository.deleteById(id);
     }
